First major patch
This is the first major patch, and it includes too many changes for me to
either recall or enumerate. I updated the help menus to no longer refer to
other programs I've written (from which I shamelessly pilfered them). I also
added several test data sets. One is completely synthetic, the other two are
real sequences, respectively a collection of reverse transcriptases and the
CEGMA subset of the KOG database.

Signed-off-by: Ted Gibbons <ted.codes@gmail.com>